+++ application/auth/src/main/java/animal/auth/presentation/UserController.java
@@ -5,9 +5,11 @@ import animal.auth.dto.UserRequest;
 import animal.auth.dto.UserResponse.UserRes;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -25,70 +27,53 @@ import response.CommonResponse.CommonEmptyRes;
 @RequestMapping("/users")
 public class UserController {
 
-    private final UserService userService;
+ /* public static final String USERNAME = "X-User-Name";
+  public static final String ROLE = "X-User-Roles";*/
 
-    /**
-     * 배송담당자 회원가입
-     */
-    @PostMapping("/delivery/sign-up")
-    public CommonResponse<CommonEmptyRes> createDeliveryUser(@Valid @RequestBody UserRequest.SignUpDeliveryReq request) {
-        userService.createDeliveryUser(request);
-        return CommonResponse.success();
-    }
+  private final UserService userService;
 
-    /**
-     * 업체 회원가입
-     */
-    @PostMapping("/company/sign-up")
-    public CommonResponse<CommonEmptyRes> createCompanyUser(@Valid @RequestBody UserRequest.SignUpCompanyReq request) {
-        userService.createCompanyUser(request);
-        return CommonResponse.success();
-    }
+  /**
+   * 사용자 목록 조회 auth : 허브 관리자 (소속된 허브 회원 리스트 조회)
+   */
+  @PreAuthorize("hasRole('MANAGER')")
+  @GetMapping("/list")
+  public CommonResponse<Page<UserRes>> getUserList(Principal principal,
+      @PageableDefault(page = 1, size = 10, sort = "createdAt", direction = Direction.DESC) Pageable pageable) {
+    var response=userService.getUserList(pageable, principal.getName());
+    return CommonResponse.success(response);
+  }
 
-    //필터 파라미터 - delivery, company
-    //회원 리스트 조회 - 해당 허브 소속의 관리자가 소속된 허브 회원리스트를 조회
-    //todo : 게이트 웨이 생성 후 다시 보기
-
-    /**
-     * 사용자 목록 조회
-     */
-    @GetMapping
-    public CommonResponse<CommonEmptyRes> getUserList(
-        @PageableDefault(page = 1, size = 10, sort = "createdAt", direction = Direction.DESC) Pageable pageable
-    ) {
-        userService.getUserList(pageable);
-        return CommonResponse.success();
-    }
-
-    /**
-     * 사용자 상세 조회
-     */
-    @GetMapping("/{username}")
-    public CommonResponse<UserRes> getUser(@PathVariable String username) {
-        var response = userService.getUserInfo(username);
-        return CommonResponse.success(response);
-    }
+  /**
+   * 사용자 상세 조회 auth : 사용자 본인, 마스터 관리자
+   */
+  @PreAuthorize("hasRole('MASTER') or @securityPermission.isResorceOwner(authentication,#username)")
+  @GetMapping("/{username}")
+  public CommonResponse<UserRes> getUser(@PathVariable String username) {
+    var response = userService.getUserInfo(username);
+    return CommonResponse.success(response);
+  }
 
 
-    /**
-     * 사용자 정보 수정
-     */
-    @PatchMapping("/{username}")
-    public CommonResponse<CommonEmptyRes> modifyCompanyUser(@PathVariable String username,
-        @Valid @RequestBody UserRequest.ModifyUserReq request) {
-        userService.modifyUser(username, request);
-        return CommonResponse.success();
-    }
+  /**
+   * 사용자 정보 수정 auth : 마스터 관리자
+   */
+  @PreAuthorize("hasRole('MASTER')")
+  @PatchMapping("/{username}")
+  public CommonResponse<CommonEmptyRes> modifyCompanyUser(@PathVariable String username,
+      @Valid @RequestBody UserRequest.ModifyUserReq request) {
+    userService.modifyUser(username, request);
+    return CommonResponse.success();
+  }
 
-    /**
-     * 사용자 탈퇴
-     */
-    @DeleteMapping("/{username}")
-    public CommonResponse<CommonEmptyRes> deleteUser(@PathVariable String username) {
-        userService.deleteUser(username);
-        return CommonResponse.success();
-    }
-
+  /**
+   * 사용자 탈퇴 auth : 마스터 관리자
+   */
+  @PreAuthorize("hasRole('MASTER')")
+  @DeleteMapping("/{username}")
+  public CommonResponse<CommonEmptyRes> deleteUser(@PathVariable String username) {
+    userService.deleteUser(username);
+    return CommonResponse.success();
+  }
 }
